Move the hashing func to a util
